Adopt SciMLTesting v1.2 folder-based run_tests (on top of grouped-tests conversion)
FVM's Core group is a capability/shared-module driver: the test files use
`using ..FiniteVolumeMethod` and a shared `test_functions.jl` include, and Core also
runs the docs/src/literate_tutorials and literate_wyos example files (outside test/)
with file-count assertions. That does not partition into self-contained group
folders, so this uses v1.2 explicit-args mode rather than forcing folder-discovery:

  * The whole original Core driver (the `safe_include` helper + the nested
    `@testset`s, verbatim) moves into test/core_tests.jl and is passed as the
    `core =` file body. The `@safetestset` wrapper supplies `using Test`, and
    core_tests.jl's `using FiniteVolumeMethod` makes the nested-module
    `using ..FiniteVolumeMethod` in each included test file resolve exactly as before.
  * QA: top-level test/qa.jl moves into test/qa/qa.jl (the existing qa sub-env), run
    as `qa = (; env = test/qa, body = test/qa/qa.jl)`.
  * `all = ["Core"]` keeps GROUP=All to Core only (QA stays its own lane), matching
    the original `if GROUP == "QA" ... else <Core> end` routing.

Adds SciMLTesting + SafeTestsets to the main test env and the qa sub-env; drops Pkg
from the test env (only the old hand-rolled harness used it).
