+++ test/Project.toml
@@ -21,10 +21,11 @@ NaturalNeighbours = "f16ad982-4edb-46b1-8125-78e5a8b5a9e6"
 NonlinearSolve = "8913a72c-1f9b-4ce2-8d82-65094dcecaec"
 OrdinaryDiffEq = "1dea7af3-3e70-54e6-95c3-0bf5283fa5ed"
 PolygonOps = "647866c9-e3ac-4575-94e7-e3d426903924"
-Pkg = "44cfe95a-1eb2-52ea-b672-e2afdf69b78f"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
 ReferenceTests = "324d217c-45ce-50fc-942e-d289b448e8cf"
 Roots = "f2b01f46-fcfa-551c-844a-d8ac1e96c665"
+SafeTestsets = "1bc83da4-3b8d-516f-aca4-4fe02f6d838f"
+SciMLTesting = "09d9d899-5365-40a9-917a-5f67fddea283"
 SimpleGraphs = "55797a34-41de-5266-9ec1-32ac4eb504d3"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
@@ -57,6 +58,8 @@ OrdinaryDiffEq = "7.0.0"
 PolygonOps = "0.1.2"
 ReferenceTests = "0.10.6"
 Roots = "3.0.0"
+SafeTestsets = "0.1, 1"
+SciMLTesting = "1"
 SimpleGraphs = "0.8.10"
 SpecialFunctions = "2.8.0"
 StableRNGs = "1.0.4"
